requirements.txt
Flask==0.10.1
Jinja2==2.7.2
MarkupSafe==0.18
PyYAML==3.10
Werkzeug==0.9.4
itsdangerous==0.23
wsgiref==0.1.2
